fix(alert) : optimize alert delete

+++ src/main/java/com/example/alerts/interfaces/rest/AlertController.java
@@ -6,6 +6,7 @@ package com.example.alerts.interfaces.rest;
 
 import com.example.alerts.application.services.AlertService;
 import com.example.alerts.domain.model.Alert;
+import com.example.diagnosis.domain.model.Diagnosis;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.HttpStatus;
@@ -56,20 +57,26 @@ public class AlertController {
     @PutMapping("/{id}")
     @Operation(summary = "Update alert by ID")
     public ResponseEntity<Alert> updateAlert(@PathVariable String id, @RequestBody Alert alert) {
-        
+
         Optional<Alert> oAlert = alertService.getAlertById(id);
-        
+
         if (oAlert.isEmpty()) {
             return ResponseEntity.notFound().build();
         }
-       
+
         return ResponseEntity.ok(alertService.updateAlert(id, alert));
     }
 
     @DeleteMapping("/{id}")
     @Operation(summary = "Delete alert by ID")
     public ResponseEntity<Void> deleteAlert(@PathVariable String id) {
+        Optional<Alert> oAlert = alertService.getAlertById(id);
+        if (oAlert.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         alertService.deleteAlert(id);
         return ResponseEntity.noContent().build();
     }
+
 }
+
